requirements.txt
flask
flask-login
werkzeug
pandas
openpyxl
qrcode[pil]
pillow
gunicorn
psycopg2-binary
